# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Statera API (Partial Static Spec)
  version: 1.0.0
  description: >-
    Partial static OpenAPI contract for the Flask backend. It does not yet cover
    health probes, worker-health endpoints, feature flags, messages,
    notifications, categories, merchants, budgets, memorized transactions, or
    the open-banking route group.
  license:
    name: Proprietary
servers:
  - url: /
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
...
    accountId:
      name: account_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    OkEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/GenericObject'
        meta:
          $ref: '#/components/schemas/GenericObject'
      additionalProperties: true
      required:
        - ok
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
        errors:
          type: array
          items:
            type: string
        code:
          type: string
        error_code:
          type: string
        retry_after:
          type: integer
      additionalProperties: true
...
    DebtAccount:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        debt_type:
          type: string
        balance_kd:
          type: string
        minimum_payment_kd:
          type: string
        due_day:
          type: integer
          nullable: true
        apr_pct:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        is_active:
          type: boolean
      additionalProperties: true
    DebtAccountRequest:
      type: object
      properties:
        name:
          type: string
        debt_type:
          type: string
          enum:
            - credit_card
            - personal_loan
            - car_loan
            - other
        balance_kd:
          type: string
        minimum_payment_kd:
          type: string
        due_day:
          type: integer
          nullable: true
        apr_pct:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      additionalProperties: true
...
paths:
...
  /api/debt-accounts:
    get:
      tags: [Debt]
      summary: List active debt accounts.
      operationId: listDebtAccounts
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Debt account list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accounts:
                            type: array
                            items:
                              $ref: '#/components/schemas/DebtAccount'
                        additionalProperties: true
    post:
      tags: [Debt]
      summary: Create a debt account.
      operationId: createDebtAccount
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebtAccountRequest'
      responses:
        '201':
          description: Debt account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/debt-accounts/summary:
    get:
      tags: [Debt]
      summary: Return debt summary metrics.
      operationId: getDebtSummary
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Debt summary payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
  /api/debt-accounts/payoff-plan:
    get:
      tags: [Debt]
      summary: Return avalanche and snowball payoff plans.
      operationId: getDebtPayoffPlan
      security:
        - sessionCookie: []
      parameters:
        - name: monthly_payment_kd
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payoff plan payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/debt-accounts/{account_id}/update:
    post:
      tags: [Debt]
      summary: Update a debt account.
      operationId: updateDebtAccount
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/accountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebtAccountRequest'
      responses:
        '200':
          description: Debt account updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/debt-accounts/{account_id}/delete:
    post:
      tags: [Debt]
      summary: Soft-delete a debt account.
      operationId: deleteDebtAccount
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/accountId'
      responses:
        '200':
          description: Debt account archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
...